Don't store options in HashMap
Previously, SaneDevice stored the options in a map of `option.name ->
option`. This is unnecessarily complex, we have to deal with duplicated
option names. devices tend not to have so many options that a linear
scan inside SaneDevice.getOption(name) is overly burdensome, therefore
we now just use an ArrayList.