Pass connection and read timeouts to image data socket.
SANE uses separate connections for control and image data. Currently,
the control socket uses connection and read timeouts that can be
specified by the user in the `SaneSession.withRemoteSane` methods. The
image data socket uses no timeouts whatsoever. This can cause infinite
delays in the case of network issues.

This change passes the user-specified connection and read timeouts to
the data socket.