Ensure the stream is correctly positioned after an error
Currently, in SANE_NET_CONTROL_OPTION we throw an exception immediately
after reading a non-zero response status. This puts the stream in a bad
state, with many bytes of unread responses confusing future calls.

This change reads off all bytes irrespective of the status message. An
exception is finally thrown when the stream is positioned appropriately.